Add - Histogram with aplha
