+++ CrytonCoreNext/Views/AIViewerView.xaml.cs
@@ -23,14 +23,8 @@ namespace CrytonCoreNext.Views
             InitializeComponent();
             DataContext = ViewModel;
             Background.Content = new FluentWaves();
-            ViewModel.OnTabControlChanged += UpdateCompareGridMaxWidth;
-            compareGrid.SizeChanged += (s,e) => UpdateCompareGridMaxWidth();
         }
 
-        private void UpdateCompareGridMaxWidth()
-        {
-            ViewModel.ImageCompareSliderValue = (int)(compareGrid.ActualWidth / 2);
-        }
 
         private void ListViewItem_MouseEnter(object sender, System.Windows.Input.MouseEventArgs e)
         {
@@ -59,6 +53,16 @@ namespace CrytonCoreNext.Views
                 scrollViewer.LineLeft();
             }
             e.Handled = true;
+        }
+
+        private void Button_MouseEnter(object sender, System.Windows.Input.MouseEventArgs e)
+        {
+            ViewModel.ShowOriginal = true;
+        }
+
+        private void Button_MouseLeave(object sender, System.Windows.Input.MouseEventArgs e)
+        {
+            ViewModel.ShowOriginal = false;
         }
     }
 }
